createworktree: move tests
