+++ tests/unit/services/test_update.py
@@ -13,7 +13,6 @@ from cutty.services.update import cherrypick
 from cutty.services.update import continueupdate
 from cutty.services.update import resetmerge
 from cutty.services.update import skipupdate
-from cutty.util.git import createworktree
 from tests.util.files import chdir
 from tests.util.git import commit
 from tests.util.git import removefile
@@ -53,50 +52,6 @@ def repository(repositorypath: Path) -> pygit2.Repository:
 def createbranch(repository: pygit2.Repository, name: str) -> pygit2.Branch:
     """Create a branch at HEAD."""
     return repository.branches.create(name, repository.head.peel())
-
-
-def test_createworktree_creates_worktree(
-    repository: pygit2.Repository, repositorypath: Path
-) -> None:
-    """It creates a worktree."""
-    createbranch(repository, "branch")
-
-    with createworktree(repositorypath, "branch") as worktree:
-        assert (worktree / ".git").is_file()
-
-
-def test_createworktree_removes_worktree_on_exit(
-    repository: pygit2.Repository, repositorypath: Path
-) -> None:
-    """It removes the worktree on exit."""
-    createbranch(repository, "branch")
-
-    with createworktree(repositorypath, "branch") as worktree:
-        pass
-
-    assert not worktree.is_dir()
-
-
-def test_createworktree_does_checkout(
-    repository: pygit2.Repository, repositorypath: Path, path: Path
-) -> None:
-    """It checks out a working tree."""
-    updatefile(path)
-    createbranch(repository, "branch")
-
-    with createworktree(repositorypath, "branch") as worktree:
-        assert (worktree / path.name).is_file()
-
-
-def test_createworktree_no_checkout(
-    repository: pygit2.Repository, repositorypath: Path, path: Path
-) -> None:
-    """It creates a worktree without checking out the files."""
-    updatefile(path)
-    createbranch(repository, "branch")
-
-    with createworktree(repositorypath, "branch", checkout=False) as worktree:
-        assert not (worktree / path.name).is_file()
 
 
 def test_cherrypick_adds_file(
